feat: :sparkles: Implement OpenTelemetry integration and LiteBus initialization for enhanced telemetry and command/query handling
