+++ NET10/Demo.Admin/Components/Pages/ServerTime.razor
@@ -1,8 +1,10 @@
 @page "/servertime"
 
+@using System.Diagnostics
 @using System.Text.Json
 @using Demo.Application.Models
 @using RestSharp
+@using OpenTelemetry.Trace
 @inject RestClient RestClient
 @inject ISnackbar Snackbar
 
@@ -107,6 +109,8 @@
 </MudContainer>
 
 @code {
+    private static readonly ActivitySource DemoAdminActivitySource = new("Demo.Admin");
+    
     private ServerTimeResponse? _serverTimeData;
     private string? _jsonResponse;
     private bool _isLoading = false;
@@ -115,6 +119,13 @@
 
     protected override async Task OnInitializedAsync()
     {
+        // 컴포넌트 초기화 추적
+        using var activity = DemoAdminActivitySource.StartActivity("ServerTime.OnInitializedAsync");
+        activity?.SetTag("blazor.component", "ServerTime");
+        activity?.SetTag("blazor.route", "/servertime");
+        activity?.SetTag("blazor.lifecycle", "OnInitializedAsync");
+        activity?.SetTag("operation.name", "component_initialization");
+        
         await RefreshTime();
     }
 
@@ -125,13 +136,35 @@
         _errorMessage = null;
         StateHasChanged();
 
+        // OpenTelemetry Activity 생성
+        using var activity = DemoAdminActivitySource.StartActivity("ServerTime API Call");
+        activity?.SetTag("operation.name", "fetch_server_time");
+        activity?.SetTag("blazor.component", "ServerTime");
+        activity?.SetTag("blazor.route", "/servertime");
+        activity?.SetTag("blazor.action", "refresh_time");
+        activity?.SetTag("component", "blazor_client");
+        activity?.SetTag("api.endpoint", "api/serverTime");
+        activity?.SetTag("api.target", "Demo.Web");
+
         try
         {
             // RestSharp을 사용한 Demo.Web serverTime API 호출
             var request = new RestRequest("api/serverTime", Method.Get);
             request.AddHeader("Accept", "application/json");
             
+            // HTTP 요청 세부 정보 추가
+            activity?.SetTag("http.method", "GET");
+            activity?.SetTag("http.url", $"{RestClient.Options.BaseUrl}api/serverTime");
+            activity?.SetTag("http.user_agent", "Demo.Admin/1.0");
+            
+            var stopwatch = Stopwatch.StartNew();
             var response = await RestClient.ExecuteAsync(request);
+            stopwatch.Stop();
+            
+            // 응답 정보 추가
+            activity?.SetTag("http.status_code", (int)response.StatusCode);
+            activity?.SetTag("http.response_time_ms", stopwatch.ElapsedMilliseconds);
+            activity?.SetTag("http.content_length", response.Content?.Length ?? 0);
             
             if (response.IsSuccessful && !string.IsNullOrEmpty(response.Content))
             {
@@ -144,12 +177,20 @@
                 
                 _serverTimeData = JsonSerializer.Deserialize<ServerTimeResponse>(_jsonResponse, options);
                 
+                activity?.SetTag("operation.result", "success");
+                activity?.SetTag("response.data_size", response.Content.Length);
+                
                 Snackbar.Add("서버 시간을 성공적으로 불러왔습니다.", Severity.Success);
             }
             else
             {
                 _hasError = true;
                 _errorMessage = $"HTTP {(int)response.StatusCode} {response.StatusCode}: {response.ErrorMessage ?? "알 수 없는 오류"}";
+                
+                activity?.SetTag("operation.result", "http_error");
+                activity?.SetTag("error.message", _errorMessage);
+                activity?.SetStatus(ActivityStatusCode.Error, _errorMessage);
+                
                 Snackbar.Add("서버 시간을 불러오는데 실패했습니다.", Severity.Error);
             }
         }
@@ -157,6 +198,13 @@
         {
             _hasError = true;
             _errorMessage = ex.Message;
+            
+            activity?.SetTag("operation.result", "exception");
+            activity?.SetTag("error.type", ex.GetType().Name);
+            activity?.SetTag("error.message", ex.Message);
+            activity?.SetStatus(ActivityStatusCode.Error, ex.Message);
+            activity?.AddException(ex);
+            
             Snackbar.Add("네트워크 오류가 발생했습니다.", Severity.Error);
         }
         finally
